pyproject.toml
[tool.poetry]
name = "obseerver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-oauth-toolkit = "^2.0.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
djangorestframework-oauth = "^1.1.0"
"backports.zoneinfo" = "^0.2.1"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
